code cleanup: remove dso_api.lib.urls, merge with dso_api.dynamic_api.urls
