+++ src/dso_api/dynamic_api/urls.py
@@ -1,7 +1,10 @@
-from django.urls import include, path
+import sys
+from importlib import import_module, reload
+
+from django.conf import settings
+from django.urls import clear_url_caches, get_urlconf, include, path
 
 from dso_api.dynamic_api.routers import DynamicRouter
-from dso_api.lib.urls import reload_urlconf
 
 from . import views
 
@@ -36,3 +39,17 @@ def refresh_urls(router_instance):
 
     # Reload the global URLs to make this visible
     reload_urlconf()
+
+
+def reload_urlconf(urlconf_name=None):
+    if urlconf_name is None:
+        urlconf_name = get_urlconf() or settings.ROOT_URLCONF
+
+    # Reload the global top-level module
+    if urlconf_name in sys.modules:
+        reload(sys.modules[urlconf_name])
+    else:
+        import_module(urlconf_name)
+
+    # Clear the Django lru caches
+    clear_url_caches()
